+++ crates/protosearch-plugin/src/options.rs
@@ -1,5 +1,5 @@
 use protobuf::UnknownValueRef;
-use protobuf::reflect::FieldDescriptor;
+use protobuf::reflect::{FieldDescriptor, MessageDescriptor};
 
 use crate::{Result, proto};
 
@@ -23,6 +23,23 @@ pub fn get_field_options(field: &FieldDescriptor) -> Result<Option<proto::Field>
         }
     }
     Ok(if found { Some(field) } else { None })
+}
+
+pub fn get_index_options(message: &MessageDescriptor) -> Result<Option<proto::Index>> {
+    use protobuf::Message;
+    let message_proto = message.proto();
+    let unknown_fields = message_proto.options.special_fields.unknown_fields();
+    let mut index = proto::Index::new();
+    let mut found = false;
+    for (number, val) in unknown_fields.iter() {
+        if number == EXTENSION_NUMBER
+            && let UnknownValueRef::LengthDelimited(b) = val
+        {
+            index.merge_from_bytes(b)?;
+            found = true;
+        }
+    }
+    Ok(if found { Some(index) } else { None })
 }
 
 /// Return `name` if specified, otherwise the field name.